Feature flag to exclude high compute  tests
